+++ Assignment-EcommerceProject-BackEnd/src/main/java/com/assignment/springboot/entity/Role.java
@@ -1,5 +1,6 @@
 package com.assignment.springboot.entity;
 
+import com.assignment.springboot.enums.RoleName;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,7 +18,7 @@ import javax.persistence.*;
 public class Role {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private long id;
 	@Enumerated(EnumType.STRING)
 	@NaturalId
 	private RoleName name;
